requirements.txt
google-adk
litellm
google-genai
feedparser
requests
python-dotenv
openai-whisper
imageio-ffmpeg
